Copy mathtools back into the project (#9)
See https://github.com/brcrista/mathtools.

I broke this out to a separate package a while ago, but going forward it will be easiest to keep in the project.  I might add more to it as I go along, and having it outside of the source tree caused problems for `mypy`.  It also makes sense to have here because that code does most of the heavy lifting in solving the problems.

I'm going to keep https://github.com/brcrista/mathtools up as an example of how to make a package and because it seems some people are using it, but I don't plan on backporting any changes there.